moved mockSerfClient to dedicated file
