+++ agent/serf.go
@@ -150,49 +150,3 @@ type retryingClient struct {
 	client *serf.RPCClient
 	config serf.Config
 }
-
-// MockSerfClient is a mock/fake Serf Client used in testing
-type MockSerfClient struct {
-	members []serf.Member
-	coords  map[string]*coordinate.Coordinate
-}
-
-// NewMockSerfClient is a helper function used to create a mock/fake Serf Client
-// used in testing
-func NewMockSerfClient(members []serf.Member, coords map[string]*coordinate.Coordinate) (client *MockSerfClient, err error) {
-	return &MockSerfClient{
-		members: members,
-		coords:  coords,
-	}, nil
-}
-
-// Members is a function that returns the Serf member nodes
-func (c *MockSerfClient) Members() ([]serf.Member, error) {
-	return c.members, nil
-}
-
-// Stop is a NOOP function used to implement the Mock Serf Client
-func (c *MockSerfClient) Stop(serf.StreamHandle) error {
-	return nil
-}
-
-// Close is a NOOP function used to implement the Mock Serf Client
-func (c *MockSerfClient) Close() error {
-	return nil
-}
-
-// Join is a NOOP function used to implement the Mock Serf Client
-func (c *MockSerfClient) Join(peers []string, replay bool) (int, error) {
-	return 0, nil
-}
-
-// UpdateTags is a NOOP function used to implement the Mock Serf Client
-func (c *MockSerfClient) UpdateTags(tags map[string]string, delTags []string) error {
-	return nil
-}
-
-// GetCoordinate get&returns the (fake) Serf Coordinate for a specific (fake) node
-// and it's mostly used during testing
-func (c *MockSerfClient) GetCoordinate(node string) (*coordinate.Coordinate, error) {
-	return c.coords[node], nil
-}
